# Generate speedup for inference