--- a/benchmarks/microbenchmarks/benchmark_inference.py
+++ b/benchmarks/microbenchmarks/benchmark_inference.py
@@ -51,9 +51,28 @@ def run(config: BenchmarkConfig) -> BenchmarkResult:
             high_precision_dtype=config.high_precision_dtype,
             device=config.device,
         )
+        # Copy base model for quantizing
+        m_copy = deepcopy(base_model)
 
-        # Use quantize_ to apply each quantization function to the model
-        m_copy = deepcopy(base_model).eval().to(config.device)
+        # Run benchmarks
+        result = BenchmarkResult(config=config)
+
+        # Store result in model for memory profiling
+        base_model._benchmark_result = result
+
+        # Run baseline benchmarking
+        base_model = base_model.eval().to(config.device)
+        if config.use_torch_compile:
+            print("Compiling baseline model....")
+            base_model = torch.compile(
+                base_model, mode=config.torch_compile_mode, fullgraph=True
+            )
+        # Benchmark time to run an inference call for baseline model
+        print("Benchmarking baseline inference.....")
+        result.baseline_inference_time_in_ms = model_inference_time_in_ms(
+            model=base_model, input_data=input_data
+        )
+
         ao_base_config = string_to_config(
             config.quantization,
             config.sparsity,
@@ -79,22 +98,27 @@ def run(config: BenchmarkConfig) -> BenchmarkResult:
             pass  # No quantization or sparsity specified, do nothing
         else:
             print("Quantizing model....")
+            m_copy = m_copy.eval().to(config.device)
             quantize_(m_copy, ao_base_config)
 
         if config.use_torch_compile:
-            print("Compiling model....")
+            print("Compiling quantized model....")
             m_copy = torch.compile(
                 m_copy, mode=config.torch_compile_mode, fullgraph=True
             )
 
-        # Run benchmarks
-        result = BenchmarkResult(config=config)
         # Store result in model for memory profiling
         m_copy._benchmark_result = result
 
         # Benchmark time to run an inference call for quantized model
+        print("Benchmarking quantized model.....")
         result.model_inference_time_in_ms = model_inference_time_in_ms(
             model=m_copy, input_data=input_data
+        )
+
+        # Calculate speedup w.r.t. baseline
+        result.speedup = round(
+            result.baseline_inference_time_in_ms / result.model_inference_time_in_ms, 2
         )
 
         # Run profiler if enabled