### PR Enhancements:
* Added a "small sweep" mode to generate shapes with increasing powers of 2 for M, K, N dimensions, ensuring M ≤ K ≤ N. This mode is useful for generating fewer combinations for visualizations like heatmaps 
* Introduced baseline inference time measurement and speedup calculation in the benchmarking process. These metrics are now stored in the `BenchmarkResult` object and displayed in the results 

## Benchmark Results:

Run the following to generate these results:
```
python -m benchmarks.microbenchmarks.benchmark_runner --config benchmarks/microbenchmarks/test/benchmark_config.yml
```

### Results on H100

| Name              | Quantization   | Shape                               | Baseline Inference Time (ms) | Inference Time (ms) | Speedup |
|-------------------|-----------------|------------|-------------------------------------|------------------------------|---------------------|
| small_bf16_linear | float8dq-tensor | small_sweep_0 (16384, 16384, 16384) | 13.34                        | 7.72                | 1.73x   |
| small_bf16_linear | float8dq-tensor | small_sweep_1 (16384, 16384, 32768) | 26.04                        | 14.62               | 1.78x   |
| small_bf16_linear | float8dq-tensor | small_sweep_2 (16384, 16384, 65536) | 53.59                        | 29.05               | 1.84x   |
| small_bf16_linear | float8dq-tensor | small_sweep_3 (16384, 32768, 32768) | 68.94                        | 28.07               | 2.46x   |
| small_bf16_linear | float8dq-tensor | small_sweep_4 (16384, 32768, 65536) | 108.63                       | 58.7                | 1.85x   |
| small_bf16_linear | float8dq-tensor | small_sweep_5 (16384, 65536, 65536) | 215.66                       | 118.42              | 1.82x   |
| small_bf16_linear | float8dq-tensor | small_sweep_6 (32768, 32768, 32768) | 108.16                       | 57.09               | 1.89x   |
| small_bf16_linear | float8dq-tensor | small_sweep_7 (32768, 32768, 65536) | 214.74                       | 110.08              | 1.95x   |
| small_bf16_linear | float8dq-tensor | small_sweep_8 (32768, 65536, 65536) | 432.44                       | 223.46              | 1.94x   |
| small_bf16_linear | float8dq-tensor | small_sweep_9 (65536, 65536, 65536) | 870.37                       | 447.97              | 1.94x   |